+++ src/chanjo2/endpoints/overview.py
@@ -1,4 +1,3 @@
-import logging
 from os import path
 
 from fastapi import APIRouter, Depends, HTTPException, Request, status
@@ -16,8 +15,6 @@ from chanjo2.meta.handle_report_contents import (
     get_report_data,
 )
 from chanjo2.models.pydantic_models import GeneCoverage, GeneReportForm, ReportQuery
-
-LOG = logging.getLogger("uvicorn.access")
 
 
 def get_templates_path() -> str: